requirements.txt
flask
flask-sqlalchemy
flask-login
authlib
python-dotenv
psycopg2-binary
waitress
gunicorn
requests
flask-seasurf
flask-limiter